moving BrowserInfo to shared component so that it can be used again for entity viewer top bar
